Refactor repo domain
