+++ lib/bacia.ex
@@ -6,7 +6,6 @@ defmodule Bacia do
   Contexts are also responsible for managing your data, regardless
   if it comes from the database, an external API or others.
   """
-alias Bacia.Repo
 
   def model do
     quote do
@@ -23,32 +22,6 @@ alias Bacia.Repo
       alias Bacia.Repo
 
       @type changeset :: Ecto.Changeset.t()
-
-      def handle_paper_trail(result) do
-          case result do
-            {:ok, %{model: model}} ->
-              {:ok, model}
-              
-            error ->
-              error
-          end
-      end
-
-      @spec fetch(struct, non_neg_integer) :: {:error, String.t()} | {:ok, struct}
-      def fetch(model, id) do
-        fetch_by(model, id: id)
-      end
-
-      @spec fetch_by(struct, non_neg_integer, list | keyword) :: {:error, String.t()} | {:ok, struct}
-      def fetch_by(model, params, opts \\ []) do
-        case Repo.get_by(model, params, opts) do
-          nil ->
-            {:error, "Entity not found"}  
-            
-          model ->
-            {:ok, model}  
-        end
-      end
     end
   end
